Update nativealign.py
Fix bug for alignments that don't have any max_n alignments

+++ yasma/nativealign.py
@@ -293,32 +293,33 @@ def align(**params):
 
 
 		if mmap == 'over':
+			if maxn_file.is_file():
 
-			with open(maxn_file, 'r') as f:
-
-
-				while True:
-					line = f.readline().strip()
-
-					if line == '':
-						break
-
-					a = pysam.AlignedSegment()
-					a.query_name = line[1:].split()[0]
-					a.flag = 4
-					a.reference_name = '*'
-					a.reference_start = -1
-					a.is_mapped = False
-					a.query_sequence = f.readline().strip()
-
-					yield a
-
-					if suff == '.fq':
-						f.readline()
-						f.readline()
+				with open(maxn_file, 'r') as f:
 
 
-			maxn_file.unlink()
+					while True:
+						line = f.readline().strip()
+
+						if line == '':
+							break
+
+						a = pysam.AlignedSegment()
+						a.query_name = line[1:].split()[0]
+						a.flag = 4
+						a.reference_name = '*'
+						a.reference_start = -1
+						a.is_mapped = False
+						a.query_sequence = f.readline().strip()
+
+						yield a
+
+						if suff == '.fq':
+							f.readline()
+							f.readline()
+
+
+				maxn_file.unlink()
 
 
 
@@ -413,7 +414,7 @@ def align(**params):
 		sys.stdout.terminal.write(to_print + '\r')
 		if not terminal_only:
 		sys.stdout.log.write(to_print + '\r')
-		
+
 		sys.stdout.flush()
